Consolidate BrowserStack desktop and mobile tests to use one common source
