+++ .github/workflows/e2e-test.yml
@@ -243,10 +243,10 @@ jobs:
           # Create .env file with secrets (for end to end run only)
           echo "TBPRO_USERNAME=${{ secrets.TBPRO_USERNAME }}" > .env
           echo "TBPRO_PASSWORD=${{ secrets.TBPRO_PASSWORD }}" >> .env
-          # The dev OIDC spec loads credentials from tests/desktop/dev/.env
+          # The dev OIDC spec loads credentials from tests/dev/.env
           # (utils/dev/credentials.ts resolves that path), so place it there too.
           cp .env packages/send/e2e
-          mv .env packages/send/e2e/tests/desktop/dev/
+          mv .env packages/send/e2e/tests/dev/
           cd packages/send/e2e
           mise exec -- pnpm run test:e2e:ci
